Tiger: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31604 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/tiger/ogrtigerdatasource.cpp
@@ -343,7 +343,7 @@ int OGRTigerDataSource::Open( const char * pszFilename, int bTestOpen,
              candidateFileList != NULL && candidateFileList[i] != NULL; 
              i++ ) 
         {
-            int nCandidateLen = strlen(candidateFileList[i]);
+            size_t nCandidateLen = strlen(candidateFileList[i]);
 
             if( papszLimitedFileList != NULL 
                 && CSLFindString(papszLimitedFileList,
